Added semiauto derivation to circe package :)
