+++ example/src/main/scala/com/itv/bucky/example/circe/CirceUnmarshalledConsumer.scala
@@ -4,7 +4,7 @@ import cats.effect.{ExitCode, IO, IOApp, Resource}
 import com.itv.bucky.AmqpClient
 import com.typesafe.scalalogging.StrictLogging
 import com.itv.bucky._
-import com.itv.bucky.circe._
+import com.itv.bucky.circe.auto._
 import com.itv.bucky.consume.{Ack, Handler}
 import com.itv.bucky.decl._
 import com.itv.bucky.example.circe.Shared.Person
